base: resolve dead-lock in weak pointer templates
This commit eliminates the mutual interlaced taking of destruction lock,
list lock and weak pointer locks that could lead to a dead-lock situation
when a lock pointer was tried to construct while a weak object is in
destruction progress. It eliminates the destruction lock completely.
Instead, all weak pointers are invalidated and dequeued at the very
beginning of the weak object's destruction. Moreover, the object is
marked as being in destruction progrss so that newly created weak pointer
detect this, and do not try to enqueue. The list operations are
done in a way that the list lock is hold at a minimum.

Fix #1607