base-foc: fix bug in Platform_thread::pause
In Platform_thread::pause, we tell FOC via exregs to raise an artificial
exception for the targeted thread. This might not apply directly to the
targeted thread if it is for instance in IPC currently. However, the
exregs call is said to return the current "CPU state" of the thread in
its &flags parameter. Through bit 0 of this CPU state, an earlier kernel
version apparently tunneled the "activity state" of the targeted thread
to the userland. Thus, we previously did a check for this bit (for
which I did not find any documentation) after the exregs, and, if it had
a specific value, waited for the targeted thread to enter exception
state before going on.

This didn't work recently at least on PBXA9. It turned out that we did
not wait for the thread although we should have. A deeper look into the
exregs API [2] and kernel back-end [1] revealed that the final flags
value comes from message register 0 in the UTCB and that this register
is merely set to the result of the architecture-specific function
user_flags() in the kernel. On ARM, this function returns always 0 and on
x86 it returns the content of the threads EFLAGS hardware register. I
could not find any place where an "activity state" bit was added.

Long story short, we now always wait for the targeted thread to enter
exception state in Platform_thread::pause as we don't have any
information about its activity state.

[1] <FOC_DIR>/kernel/fiasco/src/kern/thread_object.cpp:596
[2] <FOC_DIR>/l4/pkg/l4re-core/l4sys/include/thread.h:835

Fixes #2594